+++ source/contentRecog/recogUi.py
@@ -1,5 +1,5 @@
 # A part of NonVisual Desktop Access (NVDA)
-# Copyright (C) 2017-2023 NV Access Limited, James Teh, Leonard de RUijter
+# Copyright (C) 2017-2025 NV Access Limited, James Teh, Leonard de Ruijter, Cyrille Bougot
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
 
@@ -10,7 +10,7 @@ and present the result to the user so they can read it with cursor keys, etc.
 NVDA scripts or GUI call the L{recognizeNavigatorObject} function with the recognizer they wish to use.
 """
 
-from typing import Optional, Union
+from typing import Optional, Union, TYPE_CHECKING
 import api
 import ui
 import screenBitmap
@@ -24,7 +24,11 @@ import textInfos
 from logHandler import log
 import queueHandler
 import core
+from scriptHandler import script
 from . import RecogImageInfo, ContentRecognizer, RecognitionResult, onRecognizeResultCallbackT
+
+if TYPE_CHECKING:
+	import inputCore
 
 
 class RecogResultNVDAObject(cursorManager.CursorManager, NVDAObjects.window.Window):
@@ -109,7 +113,7 @@ class RecogResultNVDAObject(cursorManager.CursorManager, NVDAObjects.window.Wind
 
 class RefreshableRecogResultNVDAObject(RecogResultNVDAObject, LiveText):
 	"""NVDA Object that itself is responsible for fetching the recognizition result.
-	It is also able to refresh the result at intervals whenthe recognizer supports it.
+	It is also able to refresh the result at intervals or on demand when the recognizer supports it.
 	"""
 
 	def __init__(
@@ -162,17 +166,30 @@ class RefreshableRecogResultNVDAObject(RecogResultNVDAObject, LiveText):
 	def _scheduleRecognize(self):
 		core.callLater(self.recognizer.autoRefreshInterval, self._recognize, self._onResult)
 
+	@script(
+		# Translators: Describes a command.
+		description=_("Refresh the recognition result"),
+		gesture="kb:NVDA+f5",
+	)
+	def script_refreshBuffer(self, gesture: "inputCore.InputGesture") -> None:
+		if self.recognizer.allowAutoRefresh:
+			# Translators: Reported when a manual update of a content recognition result (e.g. OCR result) is
+			# requested, but the content is already updated automatically.
+			ui.message(_("The result of content recognition is already automatically updated"))
+			return
+		core.callLater(0, self._recognize, self._onResult)
+
 	def _onResult(self, result: Union[RecognitionResult, Exception]):
 		if not self.hasFocus:
 			# The user has dismissed the recognition result.
 			return
 		if isinstance(result, Exception):
-			log.error(f"Subsequent recognition failed: {result}")
+			log.error(f"Refresh recognition failed: {result}")
 			queueHandler.queueFunction(
 				queueHandler.eventQueue,
 				ui.message,
 				# Translators: Reported when recognition (e.g. OCR) fails during automatic refresh.
-				_("Automatic refresh of recognition result failed"),
+				_("Refresh of recognition result failed"),
 			)
 			self.stopMonitoring()
 			return
@@ -182,7 +199,8 @@ class RefreshableRecogResultNVDAObject(RecogResultNVDAObject, LiveText):
 		self.selection = self.makeTextInfo(self._selection.bookmark)
 		# Tell LiveText that our text has changed.
 		self.event_textChange()
-		self._scheduleRecognize()
+		if self.recognizer.allowAutoRefresh:
+			self._scheduleRecognize()
 
 	def event_gainFocus(self):
 		super().event_gainFocus()