+++ user_docs/en/changes.md
@@ -42,6 +42,7 @@ As a consequence, announcement of first line indent is now supported for LibreOf
 This option is enabled by default, but can possibly result in shorter battery life.
 If you suspect this option is negatively impacting your battery life, you're advised to disable it. (#17649, @LeonarddeR)
 * Rate boost is now supported when using Microsoft Speech API version 5 (SAPI5) and Microsoft Speech Platform voices, which supports up to 6X speed. (#17606, @gexgd0419)
+* In a recognition result, `NVDA+f5` manually refreshes the recognized content. (#17715, @CyrilleB79)
 
 ### Changes